- Created LoaderService abstraction for Oni loader and File loader - Changed OniLoaderService to conform to LoaderService abstraction - Moved LoaderService classes to submodule loader_service - Renamed FileReference to DiskFileReference and created an abstraction called FileReference - Added FileReference subclass RemoteFileReference - Removed deprecated RLoaderStrategy - Replaced OniAPIError with FileLoadError - ViewWrapperWidget tab now controls the selected LoaderService
