--- atap_corpus_loader/controller/file_loader_strategy/FileLoadError.py
@@ -1,2 +1,0 @@
-class FileLoadError(ValueError):
-    pass
