+++ atap_corpus_loader/controller/loader_service/FileLoadError.py
@@ -1,0 +1,2 @@
+class FileLoadError(Exception):
+    pass
