+++ atap_corpus_loader/controller/loader_service/__init__.py
@@ -1,0 +1,1 @@
+from .LoaderService import LoaderService
