+++ atap_corpus_loader/controller/loader_service/file_loader_strategy/__init__.py
@@ -1,3 +1,2 @@
 from .FileLoaderStrategy import FileLoaderStrategy
 from .FileLoaderFactory import FileLoaderFactory
-from .FileLoadError import FileLoadError
