+++ atap_corpus_loader/controller/loader_service/file_loader_strategy/concrete_strategies/ODSLoaderStrategy.py
@@ -1,15 +1,15 @@
 from io import BytesIO
 
-from pandas import DataFrame, read_csv
+from pandas import DataFrame, read_excel
 
 from atap_corpus_loader.controller.data_objects import CorpusHeader, DataType
-from atap_corpus_loader.controller.file_loader_strategy.FileLoaderStrategy import FileLoaderStrategy
+from atap_corpus_loader.controller.loader_service.file_loader_strategy.FileLoaderStrategy import FileLoaderStrategy
 
 
-class TSVLoaderStrategy(FileLoaderStrategy):
+class ODSLoaderStrategy(FileLoaderStrategy):
     def get_inferred_headers(self) -> list[CorpusHeader]:
         file_buf: BytesIO = self.file_ref.get_content_buffer()
-        df: DataFrame = read_csv(file_buf, sep='\t', header=0, nrows=2)
+        df: DataFrame = read_excel(file_buf, engine='odf', nrows=2)
         headers: list[CorpusHeader] = []
         for header_name, dtype_obj in df.dtypes.items():
             dtype: DataType
@@ -24,7 +24,7 @@ class TSVLoaderStrategy(FileLoaderStrategy):
     def get_dataframe(self, headers: list[CorpusHeader]) -> DataFrame:
         included_headers: list[str] = [header.name for header in headers if header.include]
         file_buf: BytesIO = self.file_ref.get_content_buffer()
-        df: DataFrame = read_csv(file_buf, sep='\t', header=0, usecols=included_headers)
+        df: DataFrame = read_excel(file_buf, engine='odf', header=0, names=included_headers)
         dtypes_applied_df: DataFrame = FileLoaderStrategy._apply_selected_dtypes(df, headers)
 
         return dtypes_applied_df
