+++ atap_corpus_loader/controller/loader_service/file_loader_strategy/concrete_strategies/TXTLoaderStrategy.py
@@ -3,7 +3,7 @@ from io import BytesIO
 from pandas import DataFrame
 
 from atap_corpus_loader.controller.data_objects import CorpusHeader, DataType
-from atap_corpus_loader.controller.file_loader_strategy.FileLoaderStrategy import FileLoaderStrategy
+from atap_corpus_loader.controller.loader_service.file_loader_strategy.FileLoaderStrategy import FileLoaderStrategy
 
 
 class TXTLoaderStrategy(FileLoaderStrategy):
