+++ atap_corpus_loader/controller/loader_service/file_loader_strategy/concrete_strategies/XMLLoaderStrategy.py
@@ -1,13 +1,22 @@
 from io import BytesIO
+from xml.etree.ElementTree import parse, Element
 
-from docx import Document
 from pandas import DataFrame
 
-from atap_corpus_loader.controller.data_objects import CorpusHeader, DataType, FileReference
-from atap_corpus_loader.controller.file_loader_strategy.FileLoaderStrategy import FileLoaderStrategy
+from atap_corpus_loader.controller.data_objects import CorpusHeader, DataType
+from atap_corpus_loader.controller.loader_service.file_loader_strategy import FileLoaderStrategy
 
 
-class DOCXLoaderStrategy(FileLoaderStrategy):
+class XMLLoaderStrategy(FileLoaderStrategy):
+    def _extract_text(self, element: Element) -> str:
+        text = element.text or ''
+        for child in element:
+            text += self._extract_text(child)
+            if child.tail:
+                text += child.tail
+
+        return text
+
     def get_inferred_headers(self) -> list[CorpusHeader]:
         headers: list[CorpusHeader] = [
             CorpusHeader('document', DataType.TEXT, include=True),
@@ -22,8 +31,10 @@ class DOCXLoaderStrategy(FileLoaderStrategy):
         file_data = {}
         if 'document' in included_headers:
             file_buf: BytesIO = self.file_ref.get_content_buffer()
-            docx_doc = Document(file_buf)
-            document = '\n'.join([p.text for p in docx_doc.paragraphs])
+            raw_xml_tree = parse(file_buf)
+            root = raw_xml_tree.getroot()
+            document = self._extract_text(root)
+
             file_data['document'] = [document]
         if 'filename' in included_headers:
             file_data['filename'] = [self.file_ref.get_filename_no_ext()]
