+++ atap_corpus_loader/controller/loader_service/file_loader_strategy/concrete_strategies/__init__.py
@@ -5,5 +5,4 @@ from .CSVLoaderStrategy import CSVLoaderStrategy
 from .TSVLoaderStrategy import TSVLoaderStrategy
 from .XLSXLoaderStrategy import XLSXLoaderStrategy
 from .ODSLoaderStrategy import ODSLoaderStrategy
-from .RLoaderStrategy import RLoaderStrategy
 from .XMLLoaderStrategy import XMLLoaderStrategy
