+++ atap_corpus_loader/view/gui/MetaEditorWidget.py
@@ -25,7 +25,7 @@ class MetaEditorWidget(AbstractWidget):
 
         corpus_table_title = Markdown("## Corpus editor")
         corpus_table_tooltip = self.view_handler.get_tooltip('corpus_editor')
-        self.corpus_table_row: Row = Row(corpus_table_title, corpus_table_tooltip)
+        self.corpus_table_row: Row = Row(corpus_table_tooltip, corpus_table_title)
 
         meta_table_title = Markdown("## Metadata editor")
         meta_table_tooltip = self.view_handler.get_tooltip('meta_editor')
